relocate proto to correct package
