+++ nats/src/main/proto/systems/ajax/event/word/update_word_event.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package systems.ajax.response_request.word;
+package systems.ajax.event.word;
 
 import "systems/ajax/entity/word.proto";
