+++ rest/src/main/kotlin/systems/ajax/englishstudytelegrambot/kafka/config/KafkaProducerConfig.kt
@@ -1,11 +1,8 @@
 package systems.ajax.englishstudytelegrambot.kafka.config
 
-import io.confluent.kafka.serializers.protobuf.KafkaProtobufDeserializerConfig
 import io.confluent.kafka.serializers.protobuf.KafkaProtobufSerializer
 import io.confluent.kafka.serializers.protobuf.KafkaProtobufSerializerConfig
 import org.apache.kafka.clients.consumer.ConsumerConfig
-import org.apache.kafka.clients.producer.KafkaProducer
-import org.apache.kafka.clients.producer.Producer
 import org.apache.kafka.clients.producer.ProducerConfig
 import org.apache.kafka.common.serialization.StringSerializer
 import org.springframework.beans.factory.annotation.Value
@@ -13,9 +10,8 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import reactor.kafka.sender.KafkaSender
 import reactor.kafka.sender.SenderOptions
-import systems.ajax.response_request.word.UpdateWordEventOuterClass.UpdateWordEvent
+import systems.ajax.event.word.UpdateWordEventOuterClass.UpdateWordEvent
 import java.util.*
-
 
 @Configuration
 class KafkaProducerConfig(
